Version 1.1: - Update Metro stations info (Al Jadaf and Creek stations were opened), update train timings - Bug fixes on MapView - Sort Journey Routes (best one is recommended) - if more than one - Share app button - Rate app button

+++ DubaiMetro/DMBarButtonItem.m
@@ -12,7 +12,7 @@
 
 - (void)setup
 {
-    const float colorMask[6] = {222, 255, 222, 255, 222, 255};
+    const CGFloat colorMask[6] = {222, 255, 222, 255, 222, 255};
     CGImageRef imageRef = CGImageCreateWithMaskingColors([[UIImage alloc] init].CGImage, colorMask);
     UIImage *image = [UIImage imageWithCGImage:imageRef];
